+++ .eslintrc.js
@@ -278,7 +278,6 @@ module.exports = {
 					// Temporary rules until we're ready to officially deprecate the bottom margins.
 					...[
 						'BaseControl',
-						'ComboboxControl',
 						'SearchControl',
 						'SelectControl',
 					].map( ( componentName ) => ( {