+++ test/e2e/specs/site-editor/template-id-format.spec.js
@@ -86,9 +86,20 @@ test.describe( 'Template ID Format', () => {
 		} );
 		await expect( editor.canvas.getByText( secondEditText ) ).toBeVisible();
 
-		await editor.saveSiteEditorEntities( {
-			isOnlyCurrentEntityDirty: true,
+		const editorTopBar = page.getByRole( 'region', {
+			name: 'Editor top bar',
 		} );
+		const saveButton = editorTopBar.getByRole( 'button', {
+			name: 'Save',
+			exact: true,
+		} );
+		await saveButton.click();
+
+		await page
+			.getByRole( 'button', { name: 'Dismiss this notice' } )
+			.getByText( /(updated|published)\./ )
+			.first()
+			.waitFor();
 	};
 
 	test( 'should open and edit templates correctly when active_templates experiment is enabled', async ( {